library/pyproject.toml
[tool.poetry]
name = "moca"
version = "0.1.0"
description = ""
authors = ["Chris Mamon <chrisam1993@gmail.com>"]
readme = "README.md"
packages = [{include = "moca", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
chromadb = "^0.4.0"
pydantic = "==1.9.*"
pyaml = "^23.7.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
pytest = "^7.4.0"
black = "^23.7.0"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning.substitution]
folders = [
  { path = "src" }
]